fix: rename ui CommitLines dir
