+++ packages/ui/src/lib/CommitLines/Cell/Straight.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { Color, Style } from '$lib/commitLines/types';
+	import type { Color, Style } from '$lib/CommitLines/types';
 
 	interface Props {
 		color: Color;
